Add filtering option on too many dropped subreads
I found CCS reads where many of the subreads were dropped can have a
lot of consensus errors (e.g. even if coverage was 10X, if there were
10 dropped reads out of 20 total to make that happen, it’s likely to be
bad).  This option removes reads with a fraction above the threshold.
This should be a small yield loss but helps high end accuracy a lot.