Fix Indexing Errors when Applying Mutations
Scenario: Imagine you have a read that spans from positions 5 to 20 on
the template.  You apply a mutation that deletes position 3.  Although
this mutation was not in the range mapped by the read, it does change
the coordinates, as the read now maps from 4 to 19.

Previously, we did not account for this, and if a mutation did not
overlap with the interval spanned by a read it would not update the
coordinates when that mutation was applied.  This could cause some
reads to be indexed +/- the true template, and thus score completely
different mutations than others.  As a result, sub-optimal consensus
results occurred (errors).