Make Enum Size Explicit
In other libraries we assume AddReadResult::OTHER is the last element
and use it to create arrays that can be indexed with any value of
AddReadResult.  However, this is a convention that is not checked by
the compiler and is entirely dependent on some programmer remembering
how this enum used in a different library.  While not a compiler guard,
explicitly creating an end SIZE element and commenting it should guard
against future users placing another value after OTHER.