Cache The Maximum Value
This change improves the speed from ~ 4.7 seconds to ~3.6 seconds (~25%
less time) on the test case given in Bugzilla # 32220.

The optimization is to avoid a second pass through the sparse matrix
vector to find the max value for normalization, but to collect that
value during the first pass and then pass it in to the finish editing
column.