Comment Cyclic Behavior Guard
This step confused me so I wanted to add comments (it wasn’t clear if
the best scoring or simply any mutation would work).  This seems to be
a nice work around that was added, and just wanted to make it clear
that one should not spend too much time deriving the provenance of the
exact steps taken.

@dalexander @nlhepler