Re-enable access to matrices from SWIG
This entailed exposing "AbstractMatrix" to the CC2 API.
AbstractMatrix can be thought of as a minimal interface exposing just
"diagnostic" capabilities, or as a mostly-opaque data type.  (In a
language with interfaces, I would have defined a "DiagnosableMatrix"
interface.)

This minimal view of Alpha/Beta can now be fetched from Evaluator or
AbstractIntegrator.