Proposed fix for crash reported by JHarting (SE-307)
Polish() contains two methods for calculating consensus
quality values, ConsensusQualities (for general QVs) and
ConsensusQVs (for rich QV information).  We caught these
errors in only the former, more widely used case, while
LAA uses the newer code-path in the latter, which lacked
that fail-safe.  Now fixed.

Technical Risk: None, since the same try/catch idiom is
used both in ConsensusQualities and elsewhere.  In addition,
in non-pathological cases, this code reduces down to the
previous code, so no downstream effects should be observed.

Not super happy with this code, since it discards potentially
useful information (from the other reads at the problem mutation)
and potentially introduces a minor discontinuity in the results
(between those w/ and w/o the problematic read) but it's the fastest,
safest way to resolve the crash for a patch.